# src/docs/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Lava Rapido API",
    "description": "API do banco de dados",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/clientes/": {
      "get": {
        "description": "Busca todos os clientes",
        "responses": {
          "200": {
            "description": "Retorna lista de clientes",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "cliente_id": {
                    "type": "number",
                    "example": 1
                  },
                  "nome": {
                    "type": "string",
                    "example": "nome do cliente"
                  },
                  "cpf": {
                    "type": "string",
                    "example": "000.000.000-00"
                  },
                  "telefone": {
                    "type": "string",
                    "example": "8500000000"
                  },
                  "enderco": {
                    "type": "string",
                    "example": "nome do endereco"
                  },
                  "email": {
                    "type": "string",
                    "example": "email@email.com"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "description": "Cria um cliente",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "nome do cliente"
                },
                "cpf": {
                  "type": "string",
                  "example": "000.000.000-00"
                },
                "telefone": {
                  "type": "string",
                  "example": "8500000000"
                },
                "enderco": {
                  "type": "string",
                  "example": "nome do endereco"
                },
                "email": {
                  "type": "string",
                  "example": "email@email.com"
                }
              },
              "required": [
                "nome",
                "cpf",
                "telefone",
                "enderco",
                "email"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registro criado com sucesso",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Registro criado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/clientes/{id}": {
      "get": {
        "description": "Busca um cliente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna um cliente",
            "schema": {
              "type": "object",
              "properties": {
                "cliente_id": {
                  "type": "number",
                  "example": 1
                },
                "nome": {
                  "type": "string",
                  "example": "nome do cliente"
                },
                "cpf": {
                  "type": "string",
                  "example": "000.000.000-00"
                },
                "telefone": {
                  "type": "string",
                  "example": "8500000000"
                },
                "enderco": {
                  "type": "string",
                  "example": "nome do endereco"
                },
                "email": {
                  "type": "string",
                  "example": "email@email.com"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "description": "Edita um cliente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "cliente_id": {
                  "type": "number",
                  "example": 1
                },
                "nome": {
                  "type": "string",
                  "example": "nome do cliente"
                },
                "cpf": {
                  "type": "string",
                  "example": "000.000.000-00"
                },
                "telefone": {
                  "type": "string",
                  "example": "8500000000"
                },
                "enderco": {
                  "type": "string",
                  "example": "nome do endereco"
                },
                "email": {
                  "type": "string",
                  "example": "email@email.com"
                }
              },
              "required": [
                "cliente_id",
                "nome",
                "cpf",
                "telefone",
                "enderco",
                "email"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registro atualizado com sucesso",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Registro atualizado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "description": "Deleta um cliente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Registro deletado",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Registro deletado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
    "/itens/": {
      "get": {
        "description": "Busca todos os itens da ordem de servico",
        "responses": {
          "200": {
            "description": "Retorna lista de itens da ordem de servico",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "itens_os_id": {
                    "type": "number",
                    "example": 1
                  },
                  "os_id": {
                    "type": "number",
                    "example": 1
                  },
                  "servico_id": {
                    "type": "number",
                    "example": 1
                  },
                  "quantidade": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "description": "Cria um item da ordem de servico",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "os_id": {
                  "type": "number",
                  "example": 1
                },
                "servico_id": {
                  "type": "number",
                  "example": 1
                },
                "quantidade": {
                  "type": "number",
                  "example": 1
                }
              },
              "required": [
                "os_id",
                "servico_id",
                "quantidade"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registro criado com sucesso",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Registro criado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/itens/{id}": {
      "get": {
        "description": "Busca um item da ordem de servico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna um item da ordem de servico",
            "schema": {
              "type": "object",
              "properties": {
                "itens_os_id": {
                  "type": "number",
                  "example": 1
                },
                "os_id": {
                  "type": "number",
                  "example": 1
                },
                "servico_id": {
                  "type": "number",
                  "example": 1
                },
                "quantidade": {
                  "type": "number",
                  "example": 1
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "description": "Edita um item da ordem de servico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "os_id": {
                  "type": "number",
                  "example": 1
                },
                "servico_id": {
                  "type": "number",
                  "example": 1
                },
                "quantidade": {
                  "type": "number",
                  "example": 1
                }
              },
              "required": [
                "os_id",
                "servico_id",
                "quantidade"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registro atualizado com sucesso",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Registro atualizado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "description": "Deleta um item da ordem de servico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Registro deletado",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Registro deletado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}